refactor: zkwallet for individual private keys
