+++ packages/client/package.json
@@ -33,6 +33,7 @@
     "@zkopru/coordinator": "file:../coordinator",
     "@zkopru/core": "file:../core",
     "@zkopru/database": "file:../database",
+    "@zkopru/zk-wizard": "file:../zk-wizard",
     "assert": "^2.0.0",
     "browserify-zlib": "^0.2.0",
     "buffer": "^6.0.3",
