+++ packages/client/src/web.ts
@@ -1,5 +1,6 @@
 import { SomeDBConnector, IndexedDBConnector } from '@zkopru/database/dist/web'
 import ZkopruNode from './zkopru-node'
+import ZkopruWallet from './zkopru-wallet'
 import RpcClient from './rpc-client'
 import { NodeConfig } from './types'
 
@@ -8,5 +9,5 @@ export default {
   Node: function ZkopruNodeDB(config: NodeConfig, connector?: SomeDBConnector) {
     return new ZkopruNode(config, connector || IndexedDBConnector)
   },
-  // Wallet: () => {},
+  Wallet: ZkopruWallet,
 }
